Enable filter expression caching by default.
